register, valid
