+++ src/main/java/com/Linguine/controller/HomeController.java
@@ -12,7 +12,7 @@ public class HomeController {
 
     @GetMapping("/login")
     public String login() {
-        return "/members/login";
+        return "members/login";
     }
 
 }
