+++ src/main/resources/static/css/styles.css
@@ -2185,7 +2185,7 @@ progress {
   color: #6c757d;
 }
 
-.form-control, .dataTable-input {
+.form-control, .dataTable-input, .form-controlfieldError{
   display: block;
   width: 100%;
   padding: 0.375rem 0.75rem;
@@ -2203,27 +2203,27 @@ progress {
   transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
 }
 @media (prefers-reduced-motion: reduce) {
-  .form-control, .dataTable-input {
+  .form-control, .dataTable-input, .form-controlfieldError{
     transition: none;
   }
 }
-.form-control[type=file], [type=file].dataTable-input {
+.form-control[type=file], [type=file].dataTable-input, .form-controlfieldError[type=file] {
   overflow: hidden;
 }
-.form-control[type=file]:not(:disabled):not([readonly]), [type=file].dataTable-input:not(:disabled):not([readonly]) {
+.form-control[type=file]:not(:disabled):not([readonly]),.form-controlfieldError[type=file]:not(:disabled):not([readonly]), [type=file].dataTable-input:not(:disabled):not([readonly]) {
   cursor: pointer;
 }
-.form-control:focus, .dataTable-input:focus {
+.form-control:focus, .dataTable-input:focus, .form-controlfieldError:focus {
   color: #212529;
   background-color: #fff;
   border-color: #86b7fe;
   outline: 0;
   box-shadow: 0 0 0 0.25rem rgba(13, 110, 253, 0.25);
 }
-.form-control::-webkit-date-and-time-value, .dataTable-input::-webkit-date-and-time-value {
+.form-control::-webkit-date-and-time-value, .dataTable-input::-webkit-date-and-time-value, .form-controlfieldError::-webkit-date-and-time-value {
   height: 1.5em;
 }
-.form-control::-moz-placeholder, .dataTable-input::-moz-placeholder {
+.form-control::-moz-placeholder, .dataTable-input::-moz-placeholder, .form-controlfieldError::-moz-placeholder {
   color: #6c757d;
   opacity: 1;
 }
