+++ src/main/resources/templates/fragments/navigation.html
@@ -21,7 +21,7 @@
         <li><a class="dropdown-item" href="#!">Activity Log</a></li>
         <li><hr class="dropdown-divider" /></li>
         <li><a class="dropdown-item" href="/login">Login</a></li>
-<!--        <li><a class="dropdown-item" href="#!">Logout</a></li>-->
+        <li><a class="dropdown-item" href="#!">Logout</a></li>
       </ul>
     </li>
   </ul>
