+++ src/main/resources/templates/members/login.html
@@ -19,10 +19,10 @@
                                         <form action = "/login" id = "loginform">
                                             <div class="form-floating mb-3">
                                                 <input class="form-control" id="inputEmail" type="email" placeholder="name@example.com" />
-                                                <label for="inputEmail">ID</label>
+                                                <label for="inputEmail">Email</label>
                                             </div>
                                             <div class="form-floating mb-3">
-                                                <input class="form-control" id="inputPassword" type="password" id = "password" placeholder="Password" />
+                                                <input class="form-control" id="inputPassword" type="password"  placeholder="Password" />
                                                 <label for="inputPassword">Password</label>
                                             </div>
                                             <div class="form-check mb-3">
@@ -30,7 +30,7 @@
                                                 <label class="form-check-label" for="inputRememberPassword">비밀번호 기억</label>
                                             </div>
                                             <div class="d-flex align-items-center justify-content-between mt-4 mb-0">
-                                                <a class="small" href="password.html">비밀번호를 잊었나요?</a>
+                                                <a class="small" href="/password">비밀번호를 잊었나요?</a>
                                                 <button class="btn btn-primary" type="submit">로그인</button>
                                             </div>
                                         </form>
