+++ src/main/resources/templates/members/password.html
@@ -1,16 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-        <meta name="description" content="" />
-        <meta name="author" content="" />
-        <title>Password Reset - SB Admin</title>
-        <link href="../../static/css/styles.css" rel="stylesheet" />
-        <script src="https://use.fontawesome.com/releases/v6.1.0/js/all.js" crossorigin="anonymous"></script>
-    </head>
+<head th:replace="fragments/header :: header" />
     <body class="bg-primary">
+    <div th:replace="fragments/bodyHeader::bodyHeader"/>
         <div id="layoutAuthentication">
             <div id="layoutAuthentication_content">
                 <main>
@@ -56,7 +48,7 @@
                 </footer>
             </div>
         </div>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
-        <script src="../../static/js/scripts.js"></script>
+
+
     </body>
 </html>
