--- src/main/resources/templates/review.html
@@ -1,12 +1,0 @@
-<html xmlns:th="http://www.thymeleaf.org">
-<body>
-<div class="container">
-    <div>
-        <h1>Linguine</h1>
-        <p>
-            <a href="/">홈</a>
-        </p> </div>
-
-</div> <!-- /container -->
-</body>
-</html>
